# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List published posts
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPostsPayload"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostPayload"
...
  /posts/{slug}:
    get:
      operationId: getPostBySlug
      tags: [posts]
      summary: Get a published post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostBySlugPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
  /posts/id/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (owner, or editor for own post)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePostPayload"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (owner, or editor for own post)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Post interactions ──────────────────────────────────────────────────
  /posts/{id}/interactions:
    get:
      operationId: getPostInteractions
      tags: [post-interactions]
      summary: Get comments and reaction counts for a post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Interactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostInteractions"
  /posts/{id}/comments:
    post:
      operationId: createPostComment
      tags: [post-interactions]
      summary: Add a comment to a post (signed-in users only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
  /posts/{id}/comments/{commentId}:
    delete:
      operationId: deletePostComment
      tags: [post-interactions]
      summary: Delete a comment (own comment or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: commentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /posts/{id}/react:
    post:
      operationId: reactToPost
      tags: [post-interactions]
      summary: Like or dislike a post (toggle, signed-in users only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Current reaction state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactToPostPayload"

  # ── Hadiths ────────────────────────────────────────────────────────────
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List inbox messages (admin only)
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMessagesPayload"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a contact message (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessagePayload"
...
  /messages/{id}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a message with its replies (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Message with replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMessagePayload"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkMessageReadPayload"
  /messages/{id}/reply:
    post:
      operationId: replyToMessage
      tags: [messages]
      summary: Reply to a message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyToMessagePayload"

  # ── Chat ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorPayload:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PostAttachment:
      type: object
      required: [url, type]
      properties:
        url: { type: string }
        type: { type: string, enum: [image, video, file] }
        name: { type: string }
    Post:
      type: object
      required: [id, title, slug, excerpt, content, published, authorName, attachments, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/PostAttachment"
        published: { type: boolean }
        authorId: { type: ["integer", "null"] }
        authorName: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ListPostsPayload:
      type: array
      items:
        $ref: "#/components/schemas/Post"
...
    PostInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/PostAttachment"
        published: { type: boolean }
    PostUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/PostAttachment"
        published: { type: boolean }
    CreatePostPayload:
      $ref: "#/components/schemas/Post"
    GetPostBySlugPayload:
      $ref: "#/components/schemas/Post"
    UpdatePostPayload:
      $ref: "#/components/schemas/Post"

    # ── Post interactions ──────────────────────────────────────────────
    PostComment:
      type: object
      required: [id, postId, authorName, isAdmin, content, createdAt]
      properties:
        id: { type: integer }
        postId: { type: integer }
        authorClerkId: { type: ["string", "null"] }
        authorName: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        isAdmin: { type: boolean }
        content: { type: string }
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [like, dislike]
    PostInteractions:
      type: object
      required: [comments, likes, dislikes, myReaction]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/PostComment"
        likes: { type: integer }
        dislikes: { type: integer }
        myReaction:
          type: ["string", "null"]
          enum: [like, dislike, null]
    ReactToPostPayload:
      type: object
      required: [myReaction]
      properties:
        myReaction:
          type: ["string", "null"]
          enum: [like, dislike, null]

    # ── Admin with avatar ──────────────────────────────────────────────
...
    Message:
      type: object
      required: [id, senderName, senderRole, content, isRead, createdAt]
      properties:
        id: { type: integer }
        senderClerkId: { type: ["string", "null"] }
        senderName: { type: string }
        senderEmail: { type: ["string", "null"] }
        senderRole: { type: string }
        recipientClerkId: { type: ["string", "null"] }
        subject: { type: ["string", "null"] }
        content: { type: string }
        isRead: { type: boolean }
        parentId: { type: ["integer", "null"] }
        createdAt: { type: string }
    ListMessagesPayload:
      type: array
      items:
        $ref: "#/components/schemas/Message"
    MessageInput:
      type: object
      required: [senderName, content]
      properties:
        senderName: { type: string, minLength: 1 }
        senderEmail: { type: string }
        subject: { type: string }
        content: { type: string, minLength: 1 }
    SendMessagePayload:
      $ref: "#/components/schemas/Message"
...
    GetMessagePayload:
      type: object
      required: [message, replies]
      properties:
        message:
          $ref: "#/components/schemas/Message"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    MarkMessageReadPayload:
      $ref: "#/components/schemas/Message"
    ReplyInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    ReplyToMessagePayload:
      $ref: "#/components/schemas/Message"

    # ── Chat ───────────────────────────────────────────────────────────
...